# apps/api/docs/openapi.yaml
openapi: 3.0.3

info:
  title: Field Service API
  description: |
    REST API for the Field Service Management Platform.

    Authentication uses short-lived JWT access tokens and rotating opaque
    refresh tokens. Users can only access and modify jobs that belong to
    their own account.
  version: 1.0.0

servers:
  - url: http://localhost:5000
    description: Local development server

tags:
  - name: Authentication
    description: User registration, login, and session information
  - name: Jobs
    description: Job management and lifecycle operations

paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account and returns an access token and refresh token.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              name: Shariful Alam
              email: shariful@example.com
              password: password123
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                data:
                  user:
                    id: user-1727000000000
                    name: Shariful Alam
                    email: shariful@example.com
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refreshToken: v4YQ3...long-random-token
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: VALIDATION_ERROR
                  message: Invalid request data
                  details:
                    - field: email
                      message: A valid email address is required
        "409":
          description: Email address already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: EMAIL_ALREADY_EXISTS
                  message: A user with this email already exists

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Log in an existing user
      description: Authenticates a user and returns an access token and refresh token.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: shariful@example.com
              password: password123
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                data:
                  user:
                    id: user-1727000000000
                    name: Shariful Alam
                    email: shariful@example.com
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refreshToken: v4YQ3...long-random-token
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: VALIDATION_ERROR
                  message: Invalid request data
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password

  /api/auth/google:
    post:
      tags:
        - Authentication
      summary: Authenticate with Google
      description: Verifies a Google ID token and creates or reuses the application user session.
      operationId: loginWithGoogle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleAuthRequest"
            example:
              idToken: eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Google authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                data:
                  user:
                    id: user-1727000000000
                    name: Shariful Alam
                    email: shariful@example.com
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refreshToken: v4YQ3...long-random-token
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: VALIDATION_ERROR
                  message: Invalid request data
        "401":
          description: Invalid Google ID token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidGoogleToken:
                  value:
                    error:
                      code: INVALID_GOOGLE_TOKEN
                      message: Invalid Google authentication token
                accountLinkRequired:
                  value:
                    error:
                      code: ACCOUNT_LINK_REQUIRED
                      message: An account already exists with this email. Sign in using the existing account before linking Google.

  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get the currently authenticated user
      description: Returns the user associated with the supplied JWT access token.
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user returned successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/User"
              example:
                data:
                  id: user-1727000000000
                  name: Shariful Alam
                  email: shariful@example.com
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: UNAUTHORIZED
                  message: Authentication required

  /api/jobs:
    get:
      tags:
        - Jobs
      summary: List the authenticated user's jobs
      description: |
        Returns only jobs belonging to the authenticated user.

        The user is identified from the JWT bearer token. No userId query
        parameter is required.
      operationId: listJobs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Jobs returned successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Job"
              example:
                data:
                  - id: job-001
                    userId: user-1727000000000
                    title: Install office air conditioner
                    scheduledAt: "2026-09-20T10:00:00+06:00"
                    location: Rajshahi City
                    status: scheduled
                    priority: high
                    createdAt: "2026-09-04T10:00:00.000Z"
                    updatedAt: "2026-09-04T10:00:00.000Z"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: UNAUTHORIZED
                  message: Authentication required

    post:
      tags:
        - Jobs
      summary: Create a new job
      description: |
        Creates a job for the authenticated user.

        The userId is derived from the JWT token and must not be supplied
        in the request body.
      operationId: createJob
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
            example:
              title: Install office air conditioner
              scheduledAt: "2026-09-20T10:00:00+06:00"
              location: Rajshahi City
              priority: high
      responses:
        "201":
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/Job"
              example:
                data:
                  id: job-001
                  userId: user-1727000000000
                  title: Install office air conditioner
                  scheduledAt: "2026-09-20T10:00:00+06:00"
                  location: Rajshahi City
                  status: scheduled
                  priority: high
                  createdAt: "2026-09-04T10:00:00.000Z"
                  updatedAt: "2026-09-04T10:00:00.000Z"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: VALIDATION_ERROR
                  message: Invalid request data
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: UNAUTHORIZED
                  message: Authentication required

  /api/jobs/{id}:
    get:
      tags:
        - Jobs
      summary: Get a job by ID
      description: |
        Returns a single job belonging to the authenticated user.

        The userId is included in the response body. It is not a path
        parameter because ownership is determined from the JWT token.
      operationId: getJobById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Public job identifier
          schema:
            type: string
          example: job-001
      responses:
        "200":
          description: Job returned successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/Job"
              example:
                data:
                  id: job-001
                  userId: user-1727000000000
                  title: Install office air conditioner
                  scheduledAt: "2026-09-20T10:00:00+06:00"
                  location: Rajshahi City
                  status: scheduled
                  priority: high
                  createdAt: "2026-09-04T10:00:00.000Z"
                  updatedAt: "2026-09-04T10:00:00.000Z"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: UNAUTHORIZED
                  message: Authentication required
        "404":
          description: Job not found or does not belong to the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: JOB_NOT_FOUND
                  message: Job not found

  /api/jobs/{id}/status:
    patch:
      tags:
        - Jobs
      summary: Update a job's status
      description: |
        Updates the status of a job belonging to the authenticated user.

        Supported lifecycle transitions:
        - scheduled → in_progress
        - in_progress → completed
      operationId: updateJobStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Public job identifier
          schema:
            type: string
          example: job-001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobStatusRequest"
            example:
              status: in_progress
      responses:
        "200":
          description: Job status updated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/Job"
              example:
                data:
                  id: job-001
                  userId: user-1727000000000
                  title: Install office air conditioner
                  scheduledAt: "2026-09-20T10:00:00+06:00"
                  location: Rajshahi City
                  status: in_progress
                  priority: high
                  createdAt: "2026-09-04T10:00:00.000Z"
                  updatedAt: "2026-09-04T11:00:00.000Z"
        "400":
          description: Invalid request data or invalid status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validationError:
                  summary: Invalid status
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid request data
                invalidTransition:
                  summary: Invalid status transition
                  value:
                    error:
                      code: INVALID_STATUS_TRANSITION
                      message: Cannot transition job from completed to in_progress
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: UNAUTHORIZED
                  message: Authentication required
        "404":
          description: Job not found or does not belong to the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: JOB_NOT_FOUND
                  message: Job not found

  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Rotate a refresh token
      description: |
        Exchanges a valid refresh token for a new short-lived access token
        and a new refresh token.

        Refresh tokens are single-use and are rotated on every successful
        refresh. The previous refresh token is invalidated.
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
            example:
              refreshToken: v4YQ3...long-random-token
      responses:
        "200":
          description: Session refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
              example:
                data:
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refreshToken: x8K2...new-rotated-token
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Refresh session is invalid, expired, revoked, or reused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidRefreshToken:
                  value:
                    error:
                      code: INVALID_REFRESH_TOKEN
                      message: Session is no longer valid. Please sign in again.
                refreshTokenReuse:
                  value:
                    error:
                      code: REFRESH_TOKEN_REUSE_DETECTED
                      message: Refresh session reuse detected. Please sign in again.

  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Log out the current refresh session
      description: Revokes the supplied refresh session on the server.
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
            example:
              refreshToken: v4YQ3...long-random-token
      responses:
        "200":
          description: Logout completed
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - success
                    properties:
                      success:
                        type: boolean
                        example: true
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter the short-lived JWT access token returned by the register,
        login, or Google authentication endpoint.

        Refresh tokens are not bearer tokens and are sent to the refresh
        endpoint in the request body.

        Format:
        Bearer <your-access-token>

  schemas:
    User:
      type: object
      description: Public user information
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
          description: Public user identifier
          example: user-1727000000000
        name:
          type: string
          maxLength: 100
          example: Shariful Alam
        email:
          type: string
          format: email
          example: shariful@example.com

    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Shariful Alam
        email:
          type: string
          format: email
          example: shariful@example.com
        password:
          type: string
          minLength: 8
          format: password
          example: password123

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: shariful@example.com
        password:
          type: string
          format: password
          example: password123

    GoogleAuthRequest:
      type: object
      required:
        - idToken
      properties:
        idToken:
          type: string
          description: Google-issued ID token obtained from the Google Sign-In flow.
          minLength: 1
          example: eyJhbGciOiJSUzI1NiIsImtpZCI6...

    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Opaque rotating refresh token.
          minLength: 1
          example: v4YQ3...long-random-token

    TokenPair:
      type: object
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
          description: Short-lived JWT access token used to authenticate API requests.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Opaque refresh token used to obtain a new access token. Rotated after each successful refresh.
          example: v4YQ3...long-random-token

    TokenPairResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/TokenPair"

    AuthResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - user
            - accessToken
            - refreshToken
          properties:
            user:
              $ref: "#/components/schemas/User"
            accessToken:
              type: string
              description: Short-lived JWT access token used to authenticate API requests.
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            refreshToken:
              type: string
              description: Opaque refresh token used to obtain a new access token. Rotated after each successful refresh.
              example: v4YQ3...long-random-token

    Job:
      type: object
      description: A field-service job owned by a user
      required:
        - id
        - userId
        - title
        - scheduledAt
        - location
        - status
        - priority
      properties:
        id:
          type: string
          description: Public job identifier
          example: job-001
        userId:
          type: string
          description: Public identifier of the user who owns this job
          example: user-1727000000000
        title:
          type: string
          description: Job title
          example: Install office air conditioner
        scheduledAt:
          type: string
          format: date-time
          description: Scheduled date and time for the job
          example: "2026-09-20T10:00:00+06:00"
        location:
          type: string
          description: Job location
          example: Rajshahi City
        status:
          type: string
          description: Current lifecycle status
          enum:
            - scheduled
            - in_progress
            - completed
          example: scheduled
        priority:
          type: string
          description: Job priority
          enum:
            - low
            - medium
            - high
            - urgent
          example: high
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2026-09-04T10:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: "2026-09-04T10:00:00.000Z"

    CreateJobRequest:
      type: object
      required:
        - title
        - scheduledAt
        - location
        - priority
      properties:
        title:
          type: string
          example: Install office air conditioner
        scheduledAt:
          type: string
          format: date-time
          example: "2026-09-20T10:00:00+06:00"
        location:
          type: string
          example: Rajshahi City
        priority:
          type: string
          enum:
            - low
            - medium
            - high
            - urgent
          example: high

    UpdateJobStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - scheduled
            - in_progress
            - completed
          example: in_progress

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request data
            details:
              type: array
              description: Optional validation details
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: A valid email address is required